Add an optional boolean field to actions
Add an optional boolean field to an action. This field being set to true
is for commands where the user might interpret anything being written to
stdout as error, e.g. 'gofmt -l .' which sets the exit code to 0 (
properly so), however the user might consider having files to format
as an error he needs to correct.