Add flag to timeoutStep for better err handling
Added an optional flag in the config file, which lets users state
that the command behaves such that on err it returns success, but the
"err" is printed to stdout. Or we treat anything printed to stdout as
error, and nothing printed to stdout as success. To enable compatibilty
between two different steps (observant and timeout) I added the flag
here and this step will probably be the main constructor. In that
case the two flags "timeout" and "stdoutAsErr" need not be
mutually exclusive.