Change output format
Desire to show status of a step as succes/failure and then
show the error and error message. In order to access the name of
the step, needed to make the name accessible. This lead to removal of
some premature abstractions. Output format will be further modified,
however this format is good enough for now.